+++ internal/cli/views/items/note/add/textarea
