+++ internal/cli/views/items/note/add/textarea.go
@@ -1,4 +1,4 @@
-package textarea
+package noteadd
 
 import (
 	"fmt"
