+++ internal/cli/views/items/note/add/textarea_test.go
@@ -1,4 +1,4 @@
-package textarea
+package noteadd
 
 import (
 	"testing"
